+++ cmd/webhook/app/BUILD.bazel
@@ -9,20 +9,15 @@ go_library(
         "//cmd/util:go_default_library",
         "//cmd/webhook/app/options:go_default_library",
         "//internal/apis/config/webhook:go_default_library",
+        "//internal/webhook:go_default_library",
         "//pkg/logs:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/feature:go_default_library",
         "//pkg/webhook:go_default_library",
         "//pkg/webhook/authority:go_default_library",
         "//pkg/webhook/configfile:go_default_library",
-        "//pkg/webhook/handlers:go_default_library",
-        "//pkg/webhook/server:go_default_library",
-        "//pkg/webhook/server/tls:go_default_library",
-        "@com_github_go_logr_logr//:go_default_library",
         "@com_github_spf13_cobra//:go_default_library",
         "@com_github_spf13_pflag//:go_default_library",
-        "@io_k8s_client_go//kubernetes:go_default_library",
-        "@io_k8s_client_go//tools/clientcmd:go_default_library",
         "@io_k8s_component_base//cli/flag:go_default_library",
     ],
 )
