+++ pkg/webhook/admission/BUILD.bazel
@@ -13,7 +13,6 @@ go_library(
     visibility = ["//:__subpackages__"],
     deps = [
         "//pkg/webhook/handlers:go_default_library",
-        "@com_github_go_logr_logr//:go_default_library",
         "@io_k8s_api//admission/v1:go_default_library",
         "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
         "@io_k8s_apimachinery//pkg/runtime:go_default_library",
@@ -55,7 +54,6 @@ go_test(
     ],
     embed = [":go_default_library"],
     deps = [
-        "//pkg/logs/testing:go_default_library",
         "//pkg/webhook/admission/initializer:go_default_library",
         "//pkg/webhook/handlers/testdata/apis/testgroup:go_default_library",
         "//pkg/webhook/handlers/testdata/apis/testgroup/install:go_default_library",
